+++ cw01/zad1/find.h
@@ -1,8 +1,7 @@
 #ifndef LIB_FIND_H
 #define LIB_FIND_H
-#include <stddef.h>
 
-int fnd_init(size_t element_max);
+int fnd_init(unsigned int element_max);
 void fnd_free();
 
 int fnd_prepare(const char* directory, const char* filename, const char* tmp);
@@ -10,8 +9,7 @@ int fnd_get_available();
 
 int fnd_search();
 
-char* fnd_get(size_t element);
-void fnd_del(size_t element);
+char* fnd_get(unsigned int element);
+void fnd_del(unsigned int element);
 
 #endif
-
